+++ scripts/build_industry_sector_ratios.py
@@ -6,8 +6,12 @@
 Build specific energy consumption by carrier and industries.
 """
 
+import logging
+
 import pandas as pd
 from _helpers import mute_print, set_scenario_config
+
+logger = logging.getLogger(__name__)
 
 # GWh/ktoe OR MWh/toe
 toe_to_MWh = 11.630
@@ -888,18 +892,48 @@ def pulp_paper_printing():
     assert s_fec.index[0] == "Paper: Paper machine - Steam use"
     assert s_ued.index[0] == "Paper: Paper machine - Steam use"
 
-    # Efficiency changes due to biomass
+    # Efficiency changes due to biomass/hydrogen
     eff_bio = s_ued["Biomass"] / s_fec["Biomass"]
-    df.loc["biomass", sector] += s_ued["Paper: Paper machine - Steam use"] / eff_bio
+    eff_h2 = s_ued["Natural gas (incl. biogas)"] / s_fec["Natural gas (incl. biogas)"]
+    # electric efficiency is assumed to be 1
+
+    df.loc["biomass", sector] += (
+        s_ued["Paper: Paper machine - Steam use"]
+        * params["steam_biomass_fraction"]
+        / eff_bio
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Paper: Paper machine - Steam use"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Paper: Paper machine - Steam use"] * params["steam_electricity_fraction"]
+    )
 
     s_fec = idees["fec"][66:77]
     s_ued = idees["ued"][66:77]
     assert s_fec.index[0] == "Paper: Product finishing - Steam use"
     assert s_ued.index[0] == "Paper: Product finishing - Steam use"
 
-    # Efficiency changes due to biomass
+    # Efficiency changes due to biomass/hydrogen
     eff_bio = s_ued["Biomass"] / s_fec["Biomass"]
-    df.loc["biomass", sector] += s_ued["Paper: Product finishing - Steam use"] / eff_bio
+    eff_h2 = s_ued["Natural gas (incl. biogas)"] / s_fec["Natural gas (incl. biogas)"]
+
+    df.loc["biomass", sector] += (
+        s_ued["Paper: Product finishing - Steam use"]
+        * params["steam_biomass_fraction"]
+        / eff_bio
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Paper: Product finishing - Steam use"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Paper: Product finishing - Steam use"]
+        * params["steam_electricity_fraction"]
+    )
 
     s_out = idees["out"][9:10]
     assert sector in str(s_out.index)
@@ -984,8 +1018,16 @@ def food_beverages_tobacco():
         s_ued["Food: Process cooling and refrigeration"] / eff_elec
     )
 
-    # Steam processing goes all to biomass without change in efficiency
-    df.loc["biomass", sector] += s_fec["Food: Steam processing"]
+    # Steam processing goes all to biomass/hydrogen/electricity without change in efficiency
+    df.loc["biomass", sector] += (
+        s_fec["Food: Steam processing"] * params["steam_biomass_fraction"]
+    )
+    df.loc["hydrogen", sector] += (
+        s_fec["Food: Steam processing"] * params["steam_hydrogen_fraction"]
+    )
+    df.loc["elec", sector] += (
+        s_fec["Food: Steam processing"] * params["steam_electricity_fraction"]
+    )
 
     # add electricity from process that is already electrified
     df.loc["elec", sector] += s_fec["Food: Electric machinery"]
@@ -1238,9 +1280,26 @@ def transport_equipment():
     df.loc["elec", sector] += s_fec["Trans. Eq.: General machinery"]
     df.loc["elec", sector] += s_fec["Trans. Eq.: Product finishing"]
 
-    # Steam processing is supplied with biomass
+    # Steam processing is supplied with biomass/hydrogen/electricity
     eff_biomass = s_ued["Biomass"] / s_fec["Biomass"]
-    df.loc["biomass", sector] += s_ued["Trans. Eq.: Steam processing"] / eff_biomass
+    eff_h2 = (
+        s_ued["Natural gas (incl. biogas)"].iloc[3]
+        / s_fec["Natural gas (incl. biogas)"].iloc[3]
+    )
+
+    df.loc["biomass", sector] += (
+        s_ued["Trans. Eq.: Steam processing"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Trans. Eq.: Steam processing"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Trans. Eq.: Steam processing"] * params["steam_electricity_fraction"]
+    )
 
     s_out = idees["out"][3:4]
     assert "Physical output" in str(s_out.index)
@@ -1290,9 +1349,26 @@ def machinery_equipment():
     df.loc["elec", sector] += s_fec["Mach. Eq.: General machinery"]
     df.loc["elec", sector] += s_fec["Mach. Eq.: Product finishing"]
 
-    # Steam processing is supplied with biomass
+    # Steam processing is supplied with biomass/hydrogen/electricity
     eff_biomass = s_ued["Biomass"] / s_fec["Biomass"]
-    df.loc["biomass", sector] += s_ued["Mach. Eq.: Steam processing"] / eff_biomass
+    eff_h2 = (
+        s_ued["Natural gas (incl. biogas)"].iloc[3]
+        / s_fec["Natural gas (incl. biogas)"].iloc[3]
+    )
+
+    df.loc["biomass", sector] += (
+        s_ued["Mach. Eq.: Steam processing"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Mach. Eq.: Steam processing"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Mach. Eq.: Steam processing"] * params["steam_electricity_fraction"]
+    )
 
     s_out = idees["out"][3:4]
     assert "Physical output" in str(s_out.index)
@@ -1333,13 +1409,41 @@ def textiles_and_leather():
     df.loc["elec", sector] += s_fec["Textiles: Electric general machinery"]
     df.loc["elec", sector] += s_fec["Textiles: Finishing Electric"]
 
-    # Steam processing is supplied with biomass
+    # Steam processing is supplied with biomass/hydrogen/electricity
     eff_biomass = s_ued[15:26]["Biomass"] / s_fec[15:26]["Biomass"]
-    df.loc["biomass", sector] += (
-        s_ued["Textiles: Pretreatment with steam"] / eff_biomass
+    eff_h2 = (
+        s_ued[15:26]["Natural gas (incl. biogas)"]
+        / s_fec[15:26]["Natural gas (incl. biogas)"]
     )
+
     df.loc["biomass", sector] += (
-        s_ued["Textiles: Wet processing with steam"] / eff_biomass
+        s_ued["Textiles: Pretreatment with steam"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Textiles: Pretreatment with steam"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Textiles: Pretreatment with steam"]
+        * params["steam_electricity_fraction"]
+    )
+
+    df.loc["biomass", sector] += (
+        s_ued["Textiles: Wet processing with steam"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Textiles: Wet processing with steam"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Textiles: Wet processing with steam"]
+        * params["steam_electricity_fraction"]
     )
 
     s_out = idees["out"][3:4]
@@ -1381,10 +1485,26 @@ def wood_and_wood_products():
     df.loc["elec", sector] += s_fec["Wood: Electric mechanical processes"]
     df.loc["elec", sector] += s_fec["Wood: Finishing Electric"]
 
-    # Steam processing is supplied with biomass
+    # Steam processing is supplied with biomass/hydrogen/electricity
     eff_biomass = s_ued[15:25]["Biomass"] / s_fec[15:25]["Biomass"]
+    eff_h2 = (
+        s_ued[15:25]["Natural gas (incl. biogas)"]
+        / s_fec[15:25]["Natural gas (incl. biogas)"]
+    )
+
     df.loc["biomass", sector] += (
-        s_ued["Wood: Specific processes with steam"] / eff_biomass
+        s_ued["Wood: Specific processes with steam"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Wood: Specific processes with steam"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Wood: Specific processes with steam"]
+        * params["steam_electricity_fraction"]
     )
 
     s_out = idees["out"][3:4]
@@ -1441,10 +1561,26 @@ def other_industrial_sectors():
     key = "Other Industrial sectors: Electric machinery"
     df.loc["elec", sector] += s_fec[key]
 
-    # Steam processing is supplied with biomass
+    # Steam processing is supplied with biomass/hydrogen/electricity
     eff_biomass = s_ued[15:25]["Biomass"] / s_fec[15:25]["Biomass"]
+    eff_h2 = (
+        s_ued[15:25]["Natural gas (incl. biogas)"]
+        / s_fec[15:25]["Natural gas (incl. biogas)"]
+    )
+
     df.loc["biomass", sector] += (
-        s_ued["Other Industrial sectors: Steam processing"] / eff_biomass
+        s_ued["Other Industrial sectors: Steam processing"]
+        * params["steam_biomass_fraction"]
+        / eff_biomass
+    )
+    df.loc["hydrogen", sector] += (
+        s_ued["Other Industrial sectors: Steam processing"]
+        * params["steam_hydrogen_fraction"]
+        / eff_h2
+    )
+    df.loc["elec", sector] += (
+        s_ued["Other Industrial sectors: Steam processing"]
+        * params["steam_electricity_fraction"]
     )
 
     s_out = idees["out"][3:4]
@@ -1470,6 +1606,23 @@ if __name__ == "__main__":
     year = 2015
 
     params = snakemake.params.industry
+
+    if (
+        params["steam_biomass_fraction"]
+        + params["steam_hydrogen_fraction"]
+        + params["steam_electricity_fraction"]
+        != 1
+    ):
+        logger.error(
+            "Fractions for steam processing do not sum to 1! Please check config[industry][steam_biomass_fraction], config[industry][steam_hydrogen_fraction] and config[industry][steam_electricity_fraction]!"
+        )
+
+    logger.info(
+        f"Steam processing fractions set to {params['steam_biomass_fraction']} biomass, {params['steam_hydrogen_fraction']} hydrogen and {params['steam_electricity_fraction']} electricity."
+    )
+    print(
+        f"Steam processing fractions set to {params['steam_biomass_fraction']} biomass, {params['steam_hydrogen_fraction']} hydrogen and {params['steam_electricity_fraction']} electricity."
+    )
 
     df = pd.concat(
         [