+++ doc/release_notes.rst
@@ -7,8 +7,11 @@
 Release Notes
 ##########################################
 
+=======
 Upcoming Release
 ================
+
+* New config ``industry: steam_biomass_fraction``, ``industry: steam_hydrogen_fraction``, ``industry: steam_electricity_fraction`` to specify the fraction of steam produced from biomass, hydrogen, and electricity, respectively. This is used to calculate custom industry sector ratios.
 
 * Allow for more conservative waste heat usage assumptions in district heating using a scaling factor for respective link efficiencies